+++ framework/Source/CPTMutableNumericData.m
@@ -119,7 +119,7 @@
 
         if ( prod != self.numberOfSamples ) {
             [NSException raise:CPTNumericDataException
-                        format:@"Shape product (%u) does not match data size (%u)", prod, self.numberOfSamples];
+                        format:@"Shape product (%lu) does not match data size (%lu)", (unsigned long)prod, (unsigned long)self.numberOfSamples];
         }
 
         shape = [shapeArray copy];
